Some fix in * / vector operators

+++ Application/EmptySource/Source/include/Math/IntVector2.h
@@ -51,8 +51,8 @@ public:
 	HOST_DEVICE FORCEINLINE IntVector2& operator*=(const int& Value);
 	HOST_DEVICE FORCEINLINE IntVector2& operator/=(const int& Value);
 
-	HOST_DEVICE inline friend IntVector2 operator*(float Value, const IntVector2 &Vector);
-	HOST_DEVICE inline friend IntVector2 operator/(float Value, const IntVector2 &Vector);
+	HOST_DEVICE inline friend IntVector2 operator*(int Value, const IntVector2 &Vector);
+	HOST_DEVICE inline friend IntVector2 operator/(int Value, const IntVector2 &Vector);
 };
 
 #include "../Math/IntVector2.inl"
